Added LogUtil, a custom debug logger that provides convenient methods for formatted Unity logging.
The contents of the file Assets/Plugins/LogUtilSource.zip are intended to be compiled as a class library in a separate project and imported as a .dll file. This ensures that the user is still able to double click on the logged message in Unity to jump directly to the line in which the log was called.