Started a proper unit testing suite for the C# side of this project. Gameplay unit testing will need to take place inside Unity itself.
Adjusted ProductLevelValueCalculator and SumLevelValueCalculator so that their base value is reflected at level 1 instead of level 0. This is more intuitive to use, since the value calculators are considered to be disabled at level 0.

Renamed AsymptoteRatioLevelValueCalculator to AsymptoteScaleLevelValueCalculator in order to more properly reflect the fact that the value goes from 1 to a specified scale value. AsymptoteRatioLevelValueCalculator will be added at a later date, which will asymptotically go from 0 to 1.